backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-multipart==0.0.12
httpx==0.27.2
python-dotenv==1.0.1
bcrypt==4.2.0
PyJWT==2.9.0
Pillow==11.0.0
aiofiles==24.1.0
